debugger/arduino-dbg/requirements.txt
pyelftools>=0.27
pyserial>=3.5
